refector: mynotificaton SSR+hydration prefetch
